resume_questionnaire_llm_bot/requirements.txt
streamlit==1.26.0
streamlit-chat==0.0.2
langchain==0.0.271
langchain-community==0.0.5
langchain-openai==0.0.4
pypdf==3.11.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
